requirements.txt
Flask==2.0.1
pytest==5.3.2
Flask-PyMongo==2.3.0
dash==2.0.0
pandas==1.3.3
flask-paginate==0.6.0
python-dotenv==0.19.0